requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
openai>=1.6.1
requests>=2.31.0
beautifulsoup4>=4.12.0
google-search-results>=2.4.2
psutil>=5.9.0
colorlog>=6.7.0
